tweak ingest.run, outward api
