+++ dcpy/lifecycle/_cli.py
@@ -1,6 +1,6 @@
 import typer
 
-from dcpy.lifecycle.ingest import run as ingest
+from dcpy.lifecycle.ingest import _cli_wrapper_run as run_ingest
 from dcpy.lifecycle.package._cli import app as package_app
 from dcpy.lifecycle.distribute import _cli as distribute_cli
 from dcpy.lifecycle.scripts import _cli as scripts_cli
@@ -11,4 +11,4 @@ app.add_typer(distribute_cli.app, name="distribute")
 app.add_typer(scripts_cli.app, name="scripts")
 
 # while there's only one ingest command, add it directly
-app.command(name="ingest")(ingest._cli_wrapper_run)
+app.command(name="ingest")(run_ingest)
